+++ Usart/inc/CUsartStateUnusable.hpp
@@ -61,7 +61,7 @@ public:
      * read does nothing in this state
      */
     void read( USART_TypeDef *  usartId,
-               uint8_t *       data,
+               uint16_t *       data,
                uint8_t          nData )
     {
 
@@ -71,7 +71,7 @@ public:
      * read does nothing in this state
      */
     void write( USART_TypeDef * usartId,
-                uint8_t *      data,
+                uint16_t *      data,
                 uint8_t         nData )
     {
